make API explicit with new entry point
- move contents of __init__ into new file sessions.py
- rename Imbox class to Session
- add deprecated Imbox function for backwards compatibility
- add entry point imbox.login (identical to Imbox)
